Subset rows based on a comparison

Subset rows based on a comparison

Usage

slice_diffs(a, b, comp, cols = everything())

slice_diffs_a(a, comp, cols = everything())

slice_diffs_b(b, comp, cols = everything())

slice_same_a(a, comp, cols = everything())

slice_same_b(b, comp, cols = everything())

Arguments

a

The original data frame .data_a passed to the tablecompare() call which created the comparison

b

The original data frame .data_b passed to the tablecompare() call which created the comparison

comp

An object of class "tbc_comparison" (the output of a tablecompare::tablecompare() call)

cols

tidy-select. For slice_diff()/slice_same() the output will only contain rows with differing/same values in one of the supplied cols.

Value

slice_diffs() returns a data.table with rows from the comparison table .data_a having values different from .data_b, followed by the corresponding row in .data_b. Only columns contained in both tables are included.

slice_diffs_a() returns a data.table with rows from the comparison table .data_a having values different from .data_b

slice_diffs_b() returns a data.table with rows from the comparison table .data_b having values different from .data_a

slice_same_a() returns a data.table with rows from the comparison table .data_a having all values the same as in .data_b

slice_same_b() returns a data.table with rows from the comparison table .data_b having all values the same as in .data_a

Examples

comp <- tblcompare(example_df_a, example_df_b, by = car)
slice_diffs(example_df_a, example_df_b, comp = comp)
#>    table            car cyl  disp drat  hp  mpg vs    wt
#> 1:     a     Datsun 710  NA 109.0 3.85  93 22.8  1 2.320
#> 2:     b     Datsun 710  NA 108.0 3.85  93 22.8  1 2.320
#> 3:     a Hornet 4 Drive   6 259.0 3.08 110 21.4  1 3.215
#> 4:     b Hornet 4 Drive   6 258.0 3.08 110 21.4  1 3.215
#> 5:     a     Duster 360   8 360.0 3.21 245 14.3  0 3.570
#> 6:     b     Duster 360   8 360.0 3.21 245 16.3  0 3.570
#> 7:     a      Merc 240D   4 146.7 3.69  62 24.4  1 3.190
#> 8:     b      Merc 240D   4 146.7 3.69  62 26.4  1 3.190
slice_diffs_a(example_df_a, comp = comp)
#>               car  mpg cyl  disp  hp drat    wt vs am
#> 1:     Datsun 710 22.8  NA 109.0  93 3.85 2.320  1  1
#> 2: Hornet 4 Drive 21.4   6 259.0 110 3.08 3.215  1  0
#> 3:     Duster 360 14.3   8 360.0 245 3.21 3.570  0  0
#> 4:      Merc 240D 24.4   4 146.7  62 3.69 3.190  1  0
slice_same_a(example_df_a, comp = comp)
#>              car  mpg cyl  disp  hp drat    wt vs am
#> 1: Mazda RX4 Wag 21.0   6 160.0 110 3.90 2.875  0  1
#> 2:      Merc 230 22.8   4 140.8  95 3.92 3.150  1  0
#> 3:     Mazda RX4 21.0   6 160.0 110 3.90 2.620  0  1
#> 4:       Valiant 18.1   6 225.0 105 2.76 3.460  1  0
#> 5:      Merc 280 19.2   6 167.6 123 3.92 3.440  1  0