Subset rows based on a comparison

Subset rows based on a comparison

Usage

slice_diffs(a, b, comp, cols = everything())

slice_diffs_a(a, comp, cols = everything())

slice_diffs_b(b, comp, cols = everything())

slice_same_a(a, comp, cols = everything())

slice_same_b(b, comp, cols = everything())

Arguments

a

The original data frame .data_a passed to the tablecompare() call which created the comparison

b

The original data frame .data_b passed to the tablecompare() call which created the comparison

comp

An object of class "tbc_comparison" (the output of a tablecompare::tablecompare() call)

cols

tidy-select. For slice_diff()/slice_same() the output will only contain rows with differing/same values in one of the supplied cols.

Value

slice_diffs() returns a data.table with rows from the comparison table .data_a having values different from .data_b, followed by the corresponding row in .data_b. Only columns contained in both tables are included.

slice_diffs_a() returns a data.table with rows from the comparison table .data_a having values different from .data_b

slice_diffs_b() returns a data.table with rows from the comparison table .data_b having values different from .data_a

slice_same_a() returns a data.table with rows from the comparison table .data_a having all values the same as in .data_b

slice_same_b() returns a data.table with rows from the comparison table .data_b having all values the same as in .data_a

Examples

# \donttest{

comp <- tblcompare(example_df_a, example_df_b, by = car)
#> Error in tblcompare(example_df_a, example_df_b, by = car): object 'example_df_a' not found
slice_diffs(example_df_a, example_df_b, comp = comp)
#> Error in slice_diffs(example_df_a, example_df_b, comp = comp): object 'comp' not found
slice_diffs_a(example_df_a, comp = comp)
#> Error in slice_diffs_a(example_df_a, comp = comp): object 'comp' not found
slice_same_a(example_df_a, comp = comp)
#> Error in slice_same_a(example_df_a, comp = comp): object 'comp' not found
# }